upload a much lower joint count skeleton
I've got my update script to export a "Lo" version of the joints.  Doesnt include the face rig, but should be a hell of aa lot easier to ingest and work with and  hopefully we can see where the problem arises from. Ran out of time tonight so no animations for the new base yet, will update my animated fbx script to accommodate the new lo skeleton and regenerate all the animation sequences.

The namespace will be updated to be simply... JR3:thenTheTransform, going forward. Ill update the hi export in one of my next commits